+++ src/duke.txt
@@ -1,3 +1,2 @@
-1. [D][X] memes (by: Mar 15, 2023 19:00)
-2. [E][ ] smd (from: monday to: tuesday)
-3. [D][ ] sdfsf (by: May 24, 2022 14:45)
+1. [T][ ] me
+2. [T][ ] erwr
